% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_column_names.R
\name{check_column_names}
\alias{check_column_names}
\title{Check column names}
\usage{
check_column_names(df, column_identifiers)
}
\arguments{
\item{df}{A data frame (see details)}

\item{column_identifiers}{A list of parameters in the data set.  See details.}
}
\value{
This function returns the parameter list, with any factors replaced by characters.
The fuction will currently stop if an error is detected, so that the parameters causing problems can be fixed by the user.
}
\description{
This function checks to make sure the variables required by PQPQ are in the data set.
}
\details{
This is a utility function designed to make sure that each parameter included in the column identfiers is present in the data set.
The parameter \code{column_identifer} should be a list with fields \code{protein_id} identifying a single varible with the proteins/Accessions, \code{confidence} identifying a single variable containing the  confidence for each peptide identity, \code{sample_names} identifying which columns contain the quantification data for each biological sample, and \code{peptide_ids} containing the columns uniquely identifying each peptide.
}
