% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQPQHelper.R
\name{clusterdata}
\alias{clusterdata}
\title{Cluster Dataset Based on Inconsistency Values}
\usage{
clusterdata(dataset, threshold = 1)
}
\arguments{
\item{dataset}{The peptides that need to be clustered}

\item{threshold}{The limit for inconsistnecy. Defaults to 1.}
}
\description{
This function does ...stuff. For later: Actually fill in this bit.
}
