% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPP.R
\name{createPepObj}
\alias{createPepObj}
\title{Create a Pep Object for PQPQ}
\usage{
createPepObj(filename, isOrig = FALSE)
}
\arguments{
\item{filename}{The file to be used to create a pepObject, in proteinPilot form.}

\item{isOrig}{Should the original Matlab version be used? Dafaults to FALSE.}
}
\value{
Returns the results from \code{\link{pqpq}} or \code{\link{pqpq_orig}}
}
\description{
This function creates a pepObject from the provided filename, filters and normalizes it,
and then sends it to the relevant PQPQ version.
}
