% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peptides.R
\name{filter_peptides}
\alias{filter_peptides}
\title{filter peptides}
\usage{
filter_peptides(pqpq_obj, column_ids, action = c("filter", "mark"))
}
\arguments{
\item{column_ids}{a list of labeled variable names.}

\item{action}{either "filter" or "mark".  See Details.}

\item{pqpq_output}{a pqpq object}
}
\description{
Filter or mark non-correlating peptides.
}
\details{
This function takes an object of class pqpq and returns a data frame.  If \code{action} = 'filter',
peptides which do not meet the PQPQ criteria are discarded.  If \code{action} = 'mark', then a new column
\code{keep} is added which is marked TRUE for peptides retained by the algoithm.
}
