% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQPQHelper.R
\name{findAllNodesBeneath}
\alias{findAllNodesBeneath}
\title{Find All Nodes Beneath A Given Hclust Node}
\usage{
findAllNodesBeneath(RowNum, hc)
}
\arguments{
\item{RowNum}{Something - for later}

\item{hc}{The hierarchal clustering result (see the package 'cluster' for details)}
}
\description{
This recursive function finds all nodes beneath the given node and returns a complete list.
}
\keyword{cluster}
\keyword{recursive}
\keyword{tree}
