% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQPQHelper.R
\name{findInconsistency}
\alias{findInconsistency}
\title{Find Inconsistency Values Of Cluster Nodes}
\usage{
findInconsistency(hclusted)
}
\arguments{
\item{hclusted}{This is the hierarchal clustering output, an object with}
}
\description{
This function computes the inconsistency values for each node in the hierarchal clustering result from cluster::hclust(). The algorithm is the one described as being used in the Matlab function 'inconsistent'.
}
