% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQPQHelper.R
\name{findSigPeptides}
\alias{findSigPeptides}
\title{Find Peptides With Significant Correlation to Model Peptide}
\usage{
findSigPeptides(ModelPeptide, CorMat, p_values, threshold = 0.4)
}
\arguments{
\item{ModelPeptide}{A number representing which row/column in the correlation matrix represents the model peptide}

\item{CorMat}{The correlation matrix for the peptides, created by psych::cor().}

\item{p_values}{The p-values matrix, created by psych::corr.test()}

\item{threshold}{The minimum value allowed for p-values. Default is 0.4 based on recommendations by the original paper.
#@source <https://www.ncbi.nlm.nih.gov/pubmed/21734112>}
}
\description{
This function uses the correlation and p-values matrices input to determine which peptides are higly correlated with the model peptide.
}
