% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize data (median method)}
\usage{
normalize_data(data, sample_names)
}
\arguments{
\item{data}{a data frame .}

\item{sample_names}{column identfiers or indices marking which columns to normalized}
}
\description{
Normalize data by dividing by median and multiplying by mean of medians
}
\details{
This function peforms median-normalization on the requested columns of a data frame.
}
