% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_selection.R
\name{peptide_selection}
\alias{peptide_selection}
\title{Select Peptides}
\usage{
peptide_selection(input_data, column_identifiers, the_limit = 0,
  p_val = 0.4, peptide_confidence_limit = 95, do_normalization = TRUE,
  protein_subset = NULL)
}
\arguments{
\item{input_data}{A data frame or list of data frames}

\item{column_identifiers}{a list of labeled variable names.}

\item{the_limit}{"peptide sum intensity limit"}

\item{p_val}{"maximum p-value of the correlation coefficient}

\item{peptide_confidence_limit}{"high confidence limit", or the minimum peptide confidence score}

\item{do_normalization}{should the data be normalized?}
}
\value{
A list, containing the cleaned data.frame and the list of labeled variable names.
}
\description{
This function does some additional processing, and validates the peptides assigned to each protein.
}
\details{
If the data is entered as a list of data frames (1 per replicate), the data is split across replicates and labeled
}
