% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQPQHelper.R
\name{peptideWithHighestAbundance}
\alias{peptideWithHighestAbundance}
\title{Find Peptide with the Highest Abundance}
\usage{
peptideWithHighestAbundance(AbundanceOnly)
}
\arguments{
\item{AbundanceOnly}{An n-by-(m - 1) matrix including only the sample measurements for the n peptides.}
}
\description{
This function finds a given peptide/row that has the highest mean value.
}
\keyword{Abundance}
\keyword{Model}
\keyword{Peptide}
