% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQPQHelper.R
\name{proteoform_ID_Four_Plus}
\alias{proteoform_ID_Four_Plus}
\title{Identify Proteoforms from Peptides}
\usage{
proteoform_ID_Four_Plus(Model_Peptide_Candidates, AbundanceOnly, protein)
}
\arguments{
\item{Model_Peptide_Candidates}{A n-by-m matrix including n peptides to be clustered.}

\item{AbundanceOnly}{Model_Peptide_Candidates, but without rownames or labels.}

\item{protein}{The name of the given protein, to be used for naming purposes.}
}
\description{
This function calls removeCols_nNAs(), clusterdata(), and peptideWithTheHighestAbundance(). It iessentially a wrapper function to make model peptide identification easier.
}
\keyword{cluster}
\keyword{proteoform}
