% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPP.R
\name{readProteinPilot}
\alias{readProteinPilot}
\title{Read a Protein Pilot Excel File}
\usage{
readProteinPilot(filename)
}
\arguments{
\item{filename}{The name/address of the file to be read}
}
\value{
Returns a list of 3 items. \code{e_data} contains the measurements from samples,
\code{f_data} contains some groupings that are nonessential for this problem, and \code{e_meta}
contains the linking between proteins and peptides.
}
\description{
This function reads information from a file, and (if it is a Protein Pilot .xlsx file)
converts it into 3 data frames to form into a pep object.
}
