% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQPQHelper.R
\name{removeCols_nNAs}
\alias{removeCols_nNAs}
\title{Remove Columns with More than N NAs}
\usage{
removeCols_nNAs(df, n)
}
\arguments{
\item{df}{The matrix with possible missing values}

\item{n}{The upper limit of NAs allowed in a column}
}
\description{
This functionr finds the number of NAs in each column, and removes any columns with more than n missing values.
}
\keyword{missing}
\keyword{values}
