% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQPQHelper.R
\name{removeLowConf}
\alias{removeLowConf}
\title{Remove Low Confidence Peptides}
\usage{
removeLowConf(data, confPeps, limit = 95)
}
\arguments{
\item{data}{An n-by-m matrix, which will be returned with rows removed}

\item{limit}{The lowest confidence level allowed. Defaults to 95.}

\item{confpeps}{A vector of length n, with confidence values from 0 to 100.}
}
\description{
This function removes columns from a matrix based on the confidence levels of each row. If there are more than 300 rows, only the 300 most-confident are allowed.
}
