% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata2.R
\docType{data}
\name{testdata2}
\alias{testdata2}
\title{testData2.xlsx}
\format{A data frame with 2804 rows and 54 attributes.}
\source{
\url{http://www.mcponline.org/content/10/10/M111.010264/suppl/DC1}
}
\usage{
testdata2
}
\description{
This is the testdata2.xlsx data provided in the PQPQ installation software.  It's provided here to demonstrate the R version of PQPQ.
It contains the Protein Pilot output for 2804 peptides.
}
\keyword{datasets}
