% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{select_peptide_data}
\usage{
normalize_data(data, sample_names)
}
\arguments{
\item{data}{a data frame .}

\item{sample_names}{column identfiers or indices marking which columns to normalized}
}
\description{
Select correlating peptides and subset them into clusters.
}
\details{
This function peforms median-normalization on the requested columns of a data frame.
}
\keyword{proteomics}
