% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQPQ_Advanced.R
\name{pqpq}
\alias{pqpq}
\title{PQPQ With Missing Values}
\usage{
pqpq(pepDataFixed, confList)
}
\arguments{
\item{pepDataFixed}{An object of type 'pepObject' (from pmartRqc), which has been preferably normalized and filtered.
The matrix 'e_data' has n rows.}

\item{confList}{A list of n confidence values from 0 to 100, corresponding to the peptides in 'e_data'.}
}
\value{
A data frame with 3 columns: \itemize{
\item \code{Protein}: Which protein this peptide is associated with
\item \code{Peptide}: The row name of the valid peptide
\item \code{ID}: Which cluster the peptide is considered clustered to}
}
\description{
This function implements the new'n'improved version of PQPQ, with an as-yet-undecided method of
avoiding NAs in the distance matrix and allowing more peptides to remain when clustering.
}
\seealso{
\code{\link{pqpq_orig}}
}
\keyword{Missing}
\keyword{PQPQ}
\keyword{values}
