% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_pqpq_data.R
\name{preprocess_pqpq_input}
\alias{preprocess_pqpq_input}
\title{Preprocess data for entry into PQPQ software}
\usage{
preprocess_pqpq_input(df, data_type = "Protein Pilot",
  protein_subset = NULL, separate_multiple_protein_IDs = FALSE,
  column_ids = NULL)
}
\arguments{
\item{df}{A data frame (see details)}

\item{data_type}{one of 'Protein Pilot', 'Spectrum Mill', or 'Proteome Discoverer', depending on the file origin.}

\item{protein_subset}{identifiers for the proteins to be analyzed.}

\item{separate_multiple_protein_IDs}{If true, peptides assigned to multiple proteins are duplicated for each protein to which they are assigned.  Otherwise, they are analyzed as a group.}

\item{column_ids}{A list of variables identifiers.}
}
\value{
A list, containing the cleaned data.frame and the list of labeled variable names.
}
\description{
This function is used to prepare Protein Pilot data for analysis in PQPQ.
}
\details{
This function handles most of the pre-processing done in the PQPQ function including:
\itemize{
\item{If \code{column_ids} is NULL, creates a list of which column names identifying the relevant portions of the data set (confidence, protein ID, peptide ID, and samples).
This is subject to error depending on how the data was input into R and the data type.}
\item{confirms that the specified columns are in the data set.}
\item{removes peptides which do not meet PQPQ criterion. (in Protein Pilot, this includes proteins with annotation 'auto - shared MS/MS')}
\item{if separate_multiple_proteins_IDs = TRUE, assigns peptides belonging to multiple proteins to ALL listed proteins separately, not the protein group.}
\item{if a protein subset is given, filters out all other proteins.}
}
Note: this is only tested on the test output from Protein Pilot provided by Jenny Forshed in the PQPQ package, and entered into R using the \code{link{openxlsx}} package.  The formatting of data from other
}
\examples{
# the column identifiers column - minimal example
data("testdata2")

# minmal column_ids example
column_ids <- list(
   protein_id = "Accessions",
   confidence = "Conf",
   samples_names = grep("Area",names(testdata2),value = TRUE),
   peptide_ids = c("Sequence", "peptide_id")
)
}
