% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_peptide_data.R
\name{select_peptide_data}
\alias{select_peptide_data}
\title{select_peptide_data}
\usage{
select_peptide_data(X, data_matrix_name, score_limit, p_val,
  peptide_confidence_limit, cut_point = 1)
}
\arguments{
\item{X}{the input data - see details.}

\item{data_matrix_name}{the name of the data matrix with the (normalized) peak intensities}

\item{score_limit}{currently unused}

\item{p_val}{the p-value associated with the significance of the correlation between 2 peptides}

\item{cut_point}{the inconsistency cut-off for clustering peptides by correlation distance.}
}
\description{
Select correlating peptides and subset them into clusters.
}
