% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPeptideData.R
\name{select_peptide_data}
\alias{select_peptide_data}
\title{select_peptide_data}
\usage{
select_peptide_data(X, score_limit, p_val, peptide_confidence_limit,
  dend_height = 1)
}
\arguments{
\item{X}{the input data - see details.}

\item{score_limit}{currently unused}

\item{p_val}{the p-value associated with the significance of the correlation between 2 peptides}

\item{dend_height}{the cut-off for clustering peptides by correlation distance.}
}
\description{
Select correlating peptides and subset them into clusters.
}
\keyword{proteomics}
